requirements.txt
Django==3.2.4
django-cors-headers==4.0.0
django-debug-toolbar==3.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-rest-passwordreset==1.3.0
drf-yasg==1.21.5
mysqlclient==2.1.1
pandas==2.0.3
python-decouple==3.8
Pillow==10.0.0
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
pytz==2023.3
requests==2.31.0
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
sqlparse==0.4.4
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==2.0.2
PyJWT==2.6.0
python-dateutil==2.8.2
django-filter==22.1
numpy==1.24.3
pytest-django==4.5
djangorestframework-recursive==0.1.2
haversine==2.7.0
cairocffi==1.6.1
svglib==1.5.1
cairosvg==2.7.1
pyfcm==1.5.4